Flatten module structure and simplify abstractions
- Flatten `src/experimental` into `src/highlight.rs`.
- Flatten `src/grammar` into `src/grammar.rs` and `src/grammar.pest`.
- Flatten `src/ast` into `src/ast.rs`.
- Flatten `src/parser` into `src/parser.rs`.
- Flatten `src/errors` into `src/errors.rs`.
- Update tests in `lambda_tests.rs` to match standard iterator method names.
- Fix clippy warning in `src/report.rs`.
- Clean up imports and file references.
